ml_engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
prophet>=1.1.5
tensorflow>=2.14.0
websockets>=12.0
openpyxl>=3.1.0
xlrd>=2.0.0
python-multipart>=0.0.9
aiosqlite>=0.20.0
aiofiles>=23.2.0
pydantic-settings>=2.0.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0
celery>=5.3.0
redis>=5.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
